data-ingester-service/requirements.txt
requests
flask
kafka-python
python-dotenv
matplotlib
pandas
flask_cors
confluent-kafka